requirements.txt
click==7.1.2
cmake==3.18.4.post1
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysqlclient==2.0.3
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.1.5
Pillow==8.1.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
Werkzeug==1.0.1
